+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # [0.6.0] - unreleased
 ## Added
+ - an installation requirement of tqdm
  - symbolic: you can now call a ``Polynomial`` object by providing an array of ``Signomial`` objects.
  - symbolic: the function ``Signomial.shift_coordates``.
  - relaxations: polynomial SAGE constraints now use ``covers`` in a way that
@@ -15,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - coniclifts: added a hotfix to the ``__init__`` function of ``PowCone3D`` objects in CVXPY. This change
    is needed due to a change in CVXPY in late 2023. This bugfix will be ported to cvxpy in a timely way
    and so this change can be reverted in the next sageopt release.
+ - coniclifts: parallel compilation of set-membership constraints.
+ - coniclifts: optional verbose compilation (gets you a progress bar on set-membership constraints).
 ## Changed
  - coniclifts: how ``disjoint_dot`` is computed in ``coniclifts.base``.
  - interop: implementation of the ``__array_priority__`` hotfix for scipy sparse matrices.